pyproject.toml
[tool.poetry]
name = "attrsfile"
version = "0.1.0"
description = "A file mapper for attrs classes."
authors = ["Deniz Bozyigit <deniz195@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
attrs = "^20.3.0"
cattrs = "^1.1.2"
"ruamel.yaml" = "^0.16.12"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
pytest = "^6.2.1"
pytest-asyncio = "^0.14.0"
jupyter = "^1.0.0"
notebook = "^6.2.0"
Sphinx = "^3.4.3"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 119
skip-string-normalization = true
target-version = ['py36', 'py37', 'py38']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist

  # The following are specific to Black, you probably don't want those.
  | blib2to3
  | tests/data
  | profiling
)/
'''
